batch: Move realized boundary lookup into package
The realization package owns entry storage while boundary lookup over those
entries remains at the batch package root.

Merge and discard policies use boundary lookup through the realized
representation, so its flat location hides which lower-level data it
examines.

This commit moves realized boundary lookup beside entry storage while
preserving the selected import path for dependent policies.

The next source commit will move working-range mapping into realization.
